Map1: changes to nameplates and changes to enemy tank info for HUD.
Nameplates now have information of the specific tank and tier the player is using, and nameplates are now twice as big for enemy tanks.

The player list for the enemy team will now hide health and tank info. Tank info is revealed when the enemy is spotted. Health is always hidden, but death status is still shown when a player dies.


Former-commit-id: bcc7ffad8afe90634406ee5848840da4ae53a464